Refactor the code. Add error message. Add success stamp.
