+++ src/components/CombineGame/CombineGame.helpers.js
@@ -1,3 +1,5 @@
+import { MESSAGES } from './constants';
+
 export function createPuzzle() {
 	let set = new Set();
 	let arr = [];
@@ -13,21 +15,22 @@ export function createPuzzle() {
 
 export function checkSingleGuess(guess, puzzle, combo) {
 	if (guess.length !== 3 || isNaN(parseInt(guess))) {
-		return false;
+		return { isCorrect: false, message: '請選 3 個圖案' };
 	}
 	if (guess[0] === guess[1] || guess[1] === guess[2] || guess[0] === guess[2]) {
-		return false;
+		return { isCorrect: false, message: '請選 3 個不一樣的圖案' };
 	}
 	const guessCombo = getCombo(guess);
 
 	for (let i = 0; i < guessCombo.length; i++) {
 		if (combo.includes(guessCombo[i])) {
-			return false;
+			return {
+				isCorrect: false,
+				message: `${guessCombo[i]} 已被回答過`,
+			};
 		}
 	}
-	const selectedShapes = guess
-		.split('')
-		.map(char => puzzle[parseInt(char) - 1]);
+	const selectedShapes = guess.split('').map(char => puzzle[Number(char) - 1]);
 	for (let i = 0; i < 3; i++) {
 		if (
 			!checkDigit(
@@ -36,10 +39,18 @@ export function checkSingleGuess(guess, puzzle, combo) {
 				selectedShapes[2][i],
 			)
 		) {
-			return false;
+			return {
+				isCorrect: false,
+				message: getErrorMessage(
+					selectedShapes[0][i],
+					selectedShapes[1][i],
+					selectedShapes[2][i],
+					i,
+				),
+			};
 		}
 	}
-	return true;
+	return { isCorrect: true, message: '' };
 }
 
 function getCombo(guess) {
@@ -51,7 +62,7 @@ function getCombo(guess) {
 
 function dfs(input, combo, ans, record) {
 	if (combo.length === input.length) {
-		ans.push(combo.join(','));
+		ans.push(combo.join(''));
 		return;
 	}
 	for (let i = 0; i < input.length; i++) {
@@ -75,16 +86,28 @@ function checkDigit(d1, d2, d3) {
 	return false;
 }
 
+function getErrorMessage(d1, d2, d3, i) {
+	let commonCode = '';
+	if (d1 === d2 || d1 === d3) {
+		commonCode = d1;
+	} else {
+		commonCode = d2;
+	}
+
+	return `${MESSAGES[i][Number(commonCode)]}有 2 個`;
+}
+
 export function checkFinish(puzzle, guessNum) {
 	const input = [1, 2, 3, 4, 5, 6, 7, 8, 9];
 	const combos = getEveryCombo(input);
 	let correctAnsNum = 0;
 	for (let i = 0; i < combos.length; i++) {
-		if (checkSingleGuess(combos[i], puzzle, [])) {
+		if (checkSingleGuess(combos[i], puzzle, []).isCorrect) {
 			correctAnsNum += 1;
 		}
 	}
-	return correctAnsNum === guessNum;
+	const isFinished = correctAnsNum === guessNum;
+	return { isFinished, message: isFinished ? '' : '還有組合沒被找到' };
 }
 
 function getEveryCombo(input) {
