+++ src/components/CombineGame/Shape.js
@@ -2,10 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import UnstyledButton from '@/components/UnstyledButton';
 import { QUERIES } from '@/constants';
-
-const SHAPES = [Circle, Square, Triangle];
-const COLORS = ['#dc2626', '#fbbf24', '#0284c7'];
-const BACKGROUND_COLORS = ['#e4e4e7', '#a1a1aa', '#52525b'];
+import { SHAPES, COLORS, BACKGROUND_COLORS } from './constants';
 
 function Shape({ id = '021', num, isSelected, handleAddNum }) {
 	const Pattern = SHAPES[id[0]];
@@ -90,31 +87,5 @@ const Label = styled.span`
 
 	transform: translateY(var(--translateY));
 `;
-
-function Circle({ fill }) {
-	return <circle cx="100" cy="100" r="60" fill={fill} />;
-}
-
-function Square({ fill }) {
-	const size = 115;
-	const distance = (200 - size) / 2;
-	return (
-		<rect width={size} height={size} x={distance} y={distance} fill={fill} />
-	);
-}
-
-function Triangle({ fill }) {
-	const cx = 100;
-	const cy = 115;
-	const r = 75;
-	const x1 = cx - (r * Math.sqrt(3)) / 2;
-	const y1 = cy + r / 2;
-	const x2 = cx + 1;
-	const y2 = cy - r;
-	const x3 = cx + (r * Math.sqrt(3)) / 2;
-	const y3 = cy + r / 2;
-	const d = `M${x1} ${y1}L${x2} ${y2}L${x3} ${y3}H${x1}Z`;
-	return <path d={d} fill={fill} />;
-}
 
 export default React.memo(Shape);
